Clean up testing
There is no need to have an entire "dummy" app structure. The same can be achieved
with a single-file app (see https://greg.molnar.io/blog/a-single-file-rails-application/)
With that, all the directories for model tests etc. can re-appear once we actually have tests
to put inside of them.

Same for fixtures - is it really needed to have a fixture with a webhook payload which
fits in a string, and use the entire Rails fixture machinery? We are after finding out
whether our webhook ends up in the DB correctly.

Less is more.