Make validation async (#13)
Most mistakes in webhook configuration are usually related to the
configuration of the secrets for signatures. While it is good in some
situations to reject the incorrectly signed webhooks immediately, in way
more situations it will actually be the misconfigured signature that'll
be at fault. A better approach is to allow the webhooks to be respooled
for processing when the signature configuration has been rectified.

- Use `valid?` in the background job instead of the controller. Most
common configuration issue is an incorrectly specified signing secret,
or an incorrectly implemented input validation. When these happen, it is
better to allow the webhook to be reprocessed
- Use instance methods in handlers instead of class methods, as they are
shorter to define. Assume a handler module supports `.new` - with a
module using singleton methods it may return `self` from `new`.
- In the config, allow the handlers specified as strings. Module
resolution in Rails happens after the config gets loaded, because the
config may alter the Zeitwerk load paths. To allow the config to get
loaded and to allow handlers to be autoloaded using Zeitwerk, the
handler modules have to be resolved lazily. This also permits the
handlers to be reloadable, like any module under Rails' autoloading
control.
- Simplify the Rails app used in tests to be small and keep it in a
single file
- If a handler is willing to expose errors to the caller, let Rails
rescue the error and display an error page or do whatever else is
configured for Rails globally.
- Store request headers with the received webhook to allow for async
validation

Closes #12 #11

---------

Co-authored-by: Stanislav Katkov <skatkov@cheddar.me>