Manage webhook state from background job (#14)
Instead of requiring the webhook handler to manage the state of the
webhook, manage it from the background job. Atomically lock the webhook
for processing, and skip the job if the webhook has already been taken
by a different worker. This removes the need for handlers to manage the
processing state, and also enables locking - removing the need for us to
have locks on the level of the job.