Limit webhook request body size (#15)
We are effectively saving arbitrary data into the DB. This can be a
large attack vector where random senders can spam the database with data
and cause a denial of service. With background validation, this is one
of the few cases where we want to reject the payload without persisting
it.