Log webhook processing into the Rails logger (#20)
The issue with Rails error reporting is that by default it does not
print into the logs, it seems. This removes a very basic affordance of
seeing that your webhooks do not validate. Raising an error during some
kind of "validation" is something Rails does _only_ when this behavior
is requested explicitly or when a `!` method is used (such as `save!` or
`create!`).

What we want is to make it easier for the developer to see their
signature verification is misconfigured, not "more correct". Besides, a
failing validation is not an error - it is a predictable (and externally
inducible) event. For instance, if someone spams the application with
webhooks with incorrect signatures - this is not an _error_, as the
application rejects those webhooks anyway.

So: when processing, log useful data into the Rails log, and only send
through errors which get raised by the handler or "upstream" operations.